Split up storage.
