+++ src/blockchain/store/chain.rs
@@ -1,48 +1,11 @@
-use std::{error::Error, fmt::Display, sync::Arc};
+use std::sync::Arc;
 
 use crate::{
     blockchain::property::*,
     stm::{abort, atomically, StmResult, TVar},
 };
 
-#[derive(Debug)]
-pub enum StoreError {
-    DoesNotExtendChain,
-}
-
-impl Display for StoreError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{:?}", self)
-    }
-}
-
-impl Error for StoreError {}
-
-/// Block level storage operations. There will be a hierarchy of these,
-/// with the era specific implementations at the bottom.
-///
-/// In this example we pretend that we only care about the main chain.
-pub trait BlockStore<E: Era> {
-    fn max_height(&self) -> StmResult<Height>;
-    /// Extend the chain with a ranking block.
-    fn add_ranking_block(&self, b: EraRankingBlock<E>) -> StmResult<()>;
-    fn get_ranking_block_by_height(&self, h: Height) -> StmResult<Option<EraRankingBlock<E>>>;
-    fn get_ranking_block_by_hash(
-        &self,
-        h: &EraRankingBlockHash<E>,
-    ) -> StmResult<Option<EraRankingBlock<E>>>;
-
-    /// Roll back all ranking blocks from the top, all the way back to a certain height.
-    fn remove_ranking_blocks_above_height(&self, h: Height) -> StmResult<()>;
-
-    /// Add an input block referenced by a ranking block.
-    fn add_input_block_header(&self, h: EraInputBlockHeader<E>) -> StmResult<()>;
-    fn has_input_block_header(&self, h: &EraInputBlockHash<E>) -> StmResult<bool>;
-    fn get_input_block_header_by_hash(
-        &self,
-        h: &EraInputBlockHash<E>,
-    ) -> StmResult<Option<EraInputBlockHeader<E>>>;
-}
+use super::{block::BlockStore, StoreError};
 
 /// Overall chain level storage, spanning all the eras.
 pub struct ChainStore<E: Era, S> {
